@@ -1,17 +1,17 @@
 name: dash_chat
 description: The most complete Chat UI for flutter inspired by react-native-gifted-chat.
-version: 1.1.18
+version: 1.2.0
 homepage: https://github.com/fayeed/dash_chat
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ">=3.3.5"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.13.2"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_parsed_text: ^2.2.1
-  uuid: ^3.0.7
+  uuid: ^4.0.0
   intl: ^0.18.0
   transparent_image: ^2.0.0
 
