@@ -242,9 +242,10 @@ class _MessageListViewState extends State<MessageListView> {
                                                           Clipboard.setData(
                                                               ClipboardData(
                                                                   text: widget
-                                                                      .messages[
-                                                                          i]
-                                                                      .text));
+                                                                          .messages[
+                                                                              i]
+                                                                          .text ??
+                                                                      ''));
                                                           Navigator.pop(
                                                               context);
                                                         },
